requirements.txt
asgiref==3.2.3
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.2
django-filter==2.2.0
djangorestframework==3.11.0
firebase-admin==3.2.1
gcloud==0.17.0
google-api-core==1.15.0
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-cloud-core==1.1.0
google-cloud-firestore==1.6.1
google-cloud-storage==1.24.1
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
gunicorn==20.0.4
httplib2==0.15.0
idna==2.8
jws==0.1.3
Markdown==3.1.1
msgpack==0.6.2
oauth2client==3.0.0
Pillow==7.0.0
protobuf==3.11.2
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
python-jwt==2.0.1
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
